split services in mappings
